[level 1] Title: 두 정수 사이의 합, Time: 12.17 ms, Memory: 90.6 MB -BaekjoonHub

+++ 프로그래머스/1/12912. 두 정수 사이의 합/두 정수 사이의 합.java
@@ -1,14 +1,12 @@
 class Solution {
     public long solution(int a, int b) {
         long answer = 0;
-        // a > b 일 때 숫자 치환
         if(a > b) {
             int temp = a;
             a = b;
             b = temp;
         }
-        // b > a 일 때 순차적으로 더하기
-        for(int i = a; i <= b; i++) {
+        for(long i = (long)a; i <= (long)b; i++) {
             answer += i;
         }
         return answer;
